refactor: extract mermaid preview hook
